refactor: add notifications
